+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/model/OpenFapState.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.faphub.installation.button.impl.model
+
+import com.flipperdevices.faphub.installation.button.api.FapButtonConfig
+
+sealed class OpenFapState {
+    object Loading : OpenFapState()
+    object NotSupported : OpenFapState()
+    object Ready : OpenFapState()
+    data class InProgress(val config: FapButtonConfig) : OpenFapState()
+}